+++ server.js
@@ -1,92 +1,29 @@
-const express = require("express");
-const fs = require("fs");
-const https = require("https");
+const express = require('express');
+const fs = require('fs');
+const https = require('https');
+const config = require('./src/common/config');
+const expressLoader = require('./src/common/modules/express');
+const socketIoLoader = require('./src/common/modules/socket');
+
 const app = express();
-const cors = require("cors");
-const socketio = require("socket.io");
 
-app.use(cors());
-const PORT = process.env.PORT || 8085;
+const socketio = require('socket.io');
+const { START_MESSAGE } = require('./src/common/constants/express');
 
-const key = fs.readFileSync("cert.key");
-const cert = fs.readFileSync("cert.crt");
+const initServer = async () => {
+    await expressLoader(app);
 
-const server = https.createServer({ key, cert }, app);
-const io = socketio.listen(server);
+    const key = fs.readFileSync('cert.key');
+    const cert = fs.readFileSync('cert.crt');
 
-let users = {};
+    const server = https.createServer({ key, cert }, app);
+    const io = socketio.listen(server);
 
-let socketToRoom = {};
+    await socketIoLoader(io);
 
-const maximum = process.env.MAXIMUM || 9;
-
-io.on("connection", (socket) => {
-  socket.on("join_room", (data) => {
-    if (users[data.room]) {
-      const length = users[data.room].length;
-      if (length === maximum) {
-        socket.to(socket.id).emit("room_full");
-        return;
-      }
-      users[data.room].push({ id: socket.id, email: data.email });
-    } else {
-      users[data.room] = [{ id: socket.id, email: data.email }];
-    }
-    socketToRoom[socket.id] = data.room;
-
-    socket.join(data.room);
-    console.log(`[${socketToRoom[socket.id]}]: ${socket.id} enter`);
-
-    const usersInThisRoom = users[data.room].filter(
-      (user) => user.id !== socket.id
-    );
-
-    console.log(usersInThisRoom);
-
-    io.sockets.to(socket.id).emit("all_users", usersInThisRoom);
-  });
-
-  socket.on("offer", (data) => {
-    //console.log(data.sdp);
-    socket.to(data.offerReceiveID).emit("getOffer", {
-      sdp: data.sdp,
-      offerSendID: data.offerSendID,
-      offerSendEmail: data.offerSendEmail,
+    server.listen(config.port, () => {
+        console.log(START_MESSAGE);
     });
-  });
+};
 
-  socket.on("answer", (data) => {
-    //console.log(data.sdp);
-    socket
-      .to(data.answerReceiveID)
-      .emit("getAnswer", { sdp: data.sdp, answerSendID: data.answerSendID });
-  });
-
-  socket.on("candidate", (data) => {
-    //console.log(data.candidate);
-    socket.to(data.candidateReceiveID).emit("getCandidate", {
-      candidate: data.candidate,
-      candidateSendID: data.candidateSendID,
-    });
-  });
-
-  socket.on("disconnect", () => {
-    console.log(`[${socketToRoom[socket.id]}]: ${socket.id} exit`);
-    const roomID = socketToRoom[socket.id];
-    let room = users[roomID];
-    if (room) {
-      room = room.filter((user) => user.id !== socket.id);
-      users[roomID] = room;
-      if (room.length === 0) {
-        delete users[roomID];
-        return;
-      }
-    }
-    socket.to(roomID).emit("user_exit", { id: socket.id });
-    console.log(users);
-  });
-});
-
-server.listen(PORT, () => {
-  console.log(`server running on ${PORT}`);
-});
+initServer();
